requirements.txt
torch
torchvision
timm
numpy
Pillow
scikit-image
gudhi
PyYAML
